backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
fpdf2>=2.7.0
